Managed various networking bugs and added features such as highlighting selected card for host and peer

+++ Client/Network.cpp
@@ -63,7 +63,7 @@ bool Network::Initialize() {
 	}
 #endif
 	for (int i = 0; i < 64; i++) {
-		sessionId += ((unsigned long long) (rand() % 2)) << i;
+		sessionId += ((uint64_t) (rand() % 2)) << i;
 	}
 	
 	return true;
